requirements.txt
discord
python-dotenv
jurigged
redis
